yuhuhero/game_back/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.1.2
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.1
bcrypt==4.0.1
python-dotenv==1.0.0 